# dist/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "CMS",
        "description": "API for CMS management",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
...
    "paths": {
        "/paragraphs": {
            "get": {
                "tags": [
                    "Paragraphs"
                ],
                "summary": "Get all paragraphs in system",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Paragraphs"
                ],
                "summary": "Create a new paragraph",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "requestBody": {
                    "description": "Paragraph object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Paragraph"
                            }
                        }
                    }
                }
            }
        },
        "/paragraphs/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the paragraph that we want to match",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "get": {
                "tags": [
                    "Paragraphs"
                ],
                "summary": "Get a paragraph",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Paragraphs"
                ],
                "summary": "Patch a paragraph",
                "requestBody": {
                    "description": "Paragraph object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Paragraph"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Paragraphs"
                ],
                "summary": "Delete a paragraph",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/sections": {
            "get": {
                "parameters": [
                    {
                        "name": "includeComponents",
                        "in": "query",
                        "description": "Include components in sections",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "includeParagraphs",
                        "in": "query",
                        "description": "Include components in components",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "tags": [
                    "Sections"
                ],
                "summary": "Get all sections in system",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sections"
                ],
                "summary": "Create a new section",
                "requestBody": {
                    "description": "Section object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Section"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/sections/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the section that we want to match",
                    "schema": {
                        "type": "string"
                    }
                },
                {
                    "name": "includeComponents",
                    "in": "query",
                    "description": "Include components in sections",
                    "schema": {
                        "type": "boolean"
                    }
                },
                {
                    "name": "includeParagraphs",
                    "in": "query",
                    "description": "Include components in components",
                    "schema": {
                        "type": "boolean"
                    }
                }
            ],
            "get": {
                "tags": [
                    "Sections"
                ],
                "summary": "Get a section",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Sections"
                ],
                "summary": "Patch a section",
                "requestBody": {
                    "description": "Section object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Section"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sections"
                ],
                "summary": "Delete a section",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Paragraph": {
            "type": "object",
            "properties": {
                "component_id": {
                    "type": "number"
                },
                "position": {
                    "type": "number"
                },
                "header": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "href": {
                    "type": "string"
                },
                "href_text": {
                    "type": "string"
                }
            }
        },
...
        "Section": {
            "type": "object",
            "properties": {
                "header": {
                    "type": "string"
                },
                "position": {
                    "type": "number"
                },
                "is_active": {
                    "type": "boolean"
                },
                "changed_date": {
                    "type": "Date"
                }
            }
        },
...
    }
}